Dance pages online
For some reason dance pages are not properly displayed. This must be resolved in a future patch. It is also important to note that the files have not been put into the right directory yet but this makes it easier to work with for now. Files must later be transferred to the right directory (for example, all CSS files for all dance pages should not be in a general map, but in the dedicated CSS map).