Dynamic adding of venues
You can now add as much venues as you want to the page of an artist (this was previously static, and always 4 venues). In this version, this is still manually. In a later version the application needs to read the amount of venues the artist plays on and then display that amount of venues on the page.